Create expression parser.
