+++ api_server/pom.xml
@@ -34,6 +34,11 @@
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
+            <groupId>com.github.groupon.monsoon</groupId>
+            <artifactId>monsoon-expr</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
             <groupId>joda-time</groupId>
             <artifactId>joda-time</artifactId>
             <version>2.8</version>
