+++ expr/src/main/java/com/groupon/lex/metrics/timeseries/parser/MutableScope.java
@@ -1,21 +1,21 @@
 /*
  * Copyright (c) 2016, Groupon, Inc.
- * All rights reserved. 
+ * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
- * are met: 
+ * are met:
  *
  * Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer. 
+ * this list of conditions and the following disclaimer.
  *
  * Redistributions in binary form must reproduce the above copyright
  * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution. 
+ * documentation and/or other materials provided with the distribution.
  *
  * Neither the name of GROUPON nor the names of its contributors may be
  * used to endorse or promote products derived from this software without
- * specific prior written permission. 
+ * specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
  * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -29,9 +29,8 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.groupon.lex.metrics.config;
+package com.groupon.lex.metrics.timeseries.parser;
 
-import static java.util.Collections.EMPTY_MAP;
 import static java.util.Collections.unmodifiableMap;
 import java.util.HashMap;
 import java.util.Map;
@@ -40,26 +39,43 @@ import java.util.Map;
  *
  * @author ariane
  */
-public class SimpleScope implements Scope {
-    private final Map<String, Type> identifiers_;
+public class MutableScope implements Scope {
+    private final Map<String, Type> identifiers_ = new HashMap<>();
 
-    public SimpleScope() { this(EMPTY_MAP); }
+    public MutableScope() {
+    }
 
-    public SimpleScope(Scope parent) {
+    public MutableScope(Scope parent) {
         this(parent.getIdentifiers());
     }
 
-    public SimpleScope(Scope parent, Map<String, Type> identifiers) {
-        this(new HashMap<String, Type>() {{
-            putAll(parent.getIdentifiers());
-            putAll(identifiers);
-        }});
+    public MutableScope(Scope parent, Map<String, Type> identifiers) {
+        this(new HashMap<String, Type>() {
+            {
+                putAll(parent.getIdentifiers());
+                putAll(identifiers);
+            }
+        });
     }
 
-    public SimpleScope(Map<String, Type> identifiers) {
-        identifiers_ = unmodifiableMap(new HashMap<>(identifiers));
+    public MutableScope(Map<String, Type> identifiers) {
+        identifiers_.putAll(identifiers);
+    }
+
+    public void put(String identifier, Type type) {
+        identifiers_.put(identifier, type);
+    }
+
+    public void putAll(Map<String, Type> identifiers) {
+        identifiers_.putAll(identifiers);
+    }
+
+    public void remove(String identifier) {
+        identifiers_.remove(identifier);
     }
 
     @Override
-    public Map<String, Type> getIdentifiers() { return identifiers_; }
+    public Map<String, Type> getIdentifiers() {
+        return unmodifiableMap(identifiers_);
+    }
 }
