+++ expr/src/main/java/com/groupon/lex/metrics/timeseries/parser/Scope.java
@@ -1,21 +1,21 @@
 /*
  * Copyright (c) 2016, Groupon, Inc.
- * All rights reserved. 
+ * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
- * are met: 
+ * are met:
  *
  * Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer. 
+ * this list of conditions and the following disclaimer.
  *
  * Redistributions in binary form must reproduce the above copyright
  * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution. 
+ * documentation and/or other materials provided with the distribution.
  *
  * Neither the name of GROUPON nor the names of its contributors may be
  * used to endorse or promote products derived from this software without
- * specific prior written permission. 
+ * specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
  * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -29,11 +29,11 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.groupon.lex.metrics.config;
+package com.groupon.lex.metrics.timeseries.parser;
 
-import static com.groupon.lex.metrics.config.Scope.Valid.DIFFERENT_IDENTIFIER;
-import static com.groupon.lex.metrics.config.Scope.Valid.NOT_AN_IDENTIFIER;
-import static com.groupon.lex.metrics.config.Scope.Valid.VALID_IDENTIFIER;
+import static com.groupon.lex.metrics.timeseries.parser.Scope.Valid.DIFFERENT_IDENTIFIER;
+import static com.groupon.lex.metrics.timeseries.parser.Scope.Valid.NOT_AN_IDENTIFIER;
+import static com.groupon.lex.metrics.timeseries.parser.Scope.Valid.VALID_IDENTIFIER;
 import static java.lang.Boolean.FALSE;
 import java.util.Map;
 import java.util.Optional;
@@ -43,7 +43,9 @@ import java.util.Optional;
  * @author ariane
  */
 public interface Scope {
-    public static Scope empty() { return new SimpleScope(); }
+    public static Scope empty() {
+        return new SimpleScope();
+    }
 
     public static enum Type {
         GROUP,
