+++ expr/src/test/java/com/groupon/lex/metrics/timeseries/parser/ScopeTest.java
@@ -1,21 +1,21 @@
 /*
  * Copyright (c) 2016, Groupon, Inc.
- * All rights reserved. 
+ * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
- * are met: 
+ * are met:
  *
  * Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer. 
+ * this list of conditions and the following disclaimer.
  *
  * Redistributions in binary form must reproduce the above copyright
  * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution. 
+ * documentation and/or other materials provided with the distribution.
  *
  * Neither the name of GROUPON nor the names of its contributors may be
  * used to endorse or promote products derived from this software without
- * specific prior written permission. 
+ * specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
  * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -29,50 +29,31 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.groupon.lex.metrics.config;
+package com.groupon.lex.metrics.timeseries.parser;
 
-import static java.util.Collections.unmodifiableMap;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Optional;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import org.junit.Test;
 
 /**
  *
  * @author ariane
  */
-public class MutableScope implements Scope {
-    private final Map<String, Type> identifiers_ = new HashMap<>();
+public class ScopeTest {
+    @Test
+    public void empty() {
+        Scope scope = Scope.empty();
 
-    public MutableScope() {}
-
-    public MutableScope(Scope parent) {
-        this(parent.getIdentifiers());
-    }
-
-    public MutableScope(Scope parent, Map<String, Type> identifiers) {
-        this(new HashMap<String, Type>() {{
-            putAll(parent.getIdentifiers());
-            putAll(identifiers);
-        }});
-    }
-
-    public MutableScope(Map<String, Type> identifiers) {
-        identifiers_.putAll(identifiers);
-    }
-
-    public void put(String identifier, Type type) {
-        identifiers_.put(identifier, type);
-    }
-
-    public void putAll(Map<String, Type> identifiers) {
-        identifiers_.putAll(identifiers);
-    }
-
-    public void remove(String identifier) {
-        identifiers_.remove(identifier);
-    }
-
-    @Override
-    public Map<String, Type> getIdentifiers() {
-        return unmodifiableMap(identifiers_);
+        assertTrue(scope.getIdentifiers().isEmpty());
+        assertEquals(Optional.empty(), scope.getIdentifier("foo"));
+        assertFalse(scope.isIdentifier("foo"));
+        assertFalse(scope.isIdentifier("foo", Scope.Type.GROUP));
+        assertFalse(scope.isIdentifier("foo", Scope.Type.METRIC));
+        assertFalse(scope.isIdentifier("foo", Scope.Type.VALUE));
+        assertEquals(Scope.Valid.NOT_AN_IDENTIFIER, scope.isValidIdentifier("foo", Scope.Type.GROUP));
+        assertEquals(Scope.Valid.NOT_AN_IDENTIFIER, scope.isValidIdentifier("foo", Scope.Type.METRIC));
+        assertEquals(Scope.Valid.NOT_AN_IDENTIFIER, scope.isValidIdentifier("foo", Scope.Type.VALUE));
     }
 }
