--- impl/src/main/antlr4/imports/ConfigBnf.tokens
@@ -1,67 +1,0 @@
-ALERT_KW=4
-ALIAS_KW=5
-ALL_KW=6
-AS_KW=7
-BANG_LIT=8
-BEGIN_QUOTE=9
-BEGIN_SQUOTE=10
-BRACE_CLOSE_LIT=11
-BRACE_OPEN_LIT=12
-COLLECTORS_KW=13
-COLLECT_KW=14
-COLON_LIT=15
-COMMA_LIT=16
-COMMENT=17
-CONSTANT_KW=18
-CURLYBRACKET_CLOSE=19
-CURLYBRACKET_OPEN=20
-DASH_LIT=21
-DEFINE_KW=22
-DIGITS=23
-DOLLAR_LIT=24
-DOT_LIT=25
-ENDSTATEMENT_KW=26
-END_QUOTE=27
-EQ_KW=28
-ESC_CHAR=29
-FALSE_KW=30
-FOR_KW=31
-FP_DECIMAL=32
-FP_HEX=33
-FROM_KW=34
-GE_KW=35
-GT_KW=36
-HEXDIGITS=37
-ID=38
-IF_KW=39
-IMPORT_KW=40
-LEFTSHIFT_KW=41
-LE_KW=42
-LOGICAL_AND_KW=43
-LOGICAL_OR_KW=44
-LT_KW=45
-MATCH_KW=46
-MESSAGE_KW=47
-NEQ_KW=48
-OCTDIGITS=49
-PERCENT_LIT=50
-PLUS_LIT=51
-RAW=52
-REGEX_MATCH_KW=53
-REGEX_NEGATE_KW=54
-RIGHTSHIFT_KW=55
-SLASH_LIT=56
-SQBRACE_CLOSE_LIT=57
-SQBRACE_OPEN_LIT=58
-STAR=59
-STAR_STAR=60
-TAG_KW=61
-TRUE_KW=62
-WHERE_KW=63
-WS=64
-BY_KW=65
-KEEP_COMMON_KW=66
-WITHOUT_KW=67
-ATTRIBUTES_KW=68
-BEGIN_REGEX=69
-DOT_DOT_LIT=70
